move handlebars to helpers
